Refactor injectable element initialization to reuse the logic
